extract errors to file
